moved NavigationPageView and NavigationPageViewModel to the right directory, added UI test
