+++ App/automotive/src/androidTest/java/com/amos/infotaimos/NavigationPageTest.kt
@@ -14,6 +14,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.amos.infotaimos.model.CarInstanceManager
 import com.amos.infotaimos.model.NavigationService
+import com.amos.infotaimos.view.NavigationPage
 import org.hamcrest.Matchers.allOf
 import org.hamcrest.Matchers.instanceOf
 import org.hamcrest.Matchers.`is`
@@ -35,7 +36,11 @@ class NavigationPageTest {
         launchFragmentInContainer<NavigationPage>()
         onView(withId(R.id.stop_navigation_button)).perform(click()).check(matches(isClickable()))
     }
-
+    @Test
+    fun testAnnouncementClickable() {
+        launchFragmentInContainer<NavigationPage>()
+        onView(withId(R.id.simulate_speech_announcment)).perform(click()).check(matches(isClickable()))
+    }
     @Test
     fun startStopNavigation() {
         val context = ApplicationProvider.getApplicationContext<Context>()
