+++ App/automotive/src/main/java/com/amos/infotaimos/view/NavigationPage.kt
@@ -1,9 +1,11 @@
-package com.amos.infotaimos
+package com.amos.infotaimos.view
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.viewModels
+import com.amos.infotaimos.ViewBindingFragment
 import com.amos.infotaimos.databinding.FragmentNavigationPageBinding
+import com.amos.infotaimos.viewmodel.NavigationPageViewModel
 
 class NavigationPage : ViewBindingFragment<FragmentNavigationPageBinding>() {
     private val viewModel: NavigationPageViewModel by viewModels()
