+++ App/automotive/src/main/java/com/amos/infotaimos/viewmodel/NavigationPageViewModel.kt
@@ -1,4 +1,4 @@
-package com.amos.infotaimos
+package com.amos.infotaimos.viewmodel
 
 import android.content.ContentResolver
 import android.content.Context
@@ -6,6 +6,7 @@ import android.media.AudioAttributes
 import android.media.MediaPlayer
 import android.net.Uri
 import androidx.lifecycle.ViewModel
+import com.amos.infotaimos.R
 import com.amos.infotaimos.model.CarInstanceManager
 import com.amos.infotaimos.model.NavigationService
