+++ App/automotive/src/main/res/layout/fragment_navigation_page.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".NavigationPage">
+    tools:context=".view.NavigationPage">
 
     <LinearLayout
         android:layout_width="match_parent"
